fix: transfer of coponenots to Ui-kit
