+++ src/components/UI-Kit/FiltersInputSearch/FiltersInputSearch.style.ts
