+++ src/components/UI-Kit/FiltersInputSearch/index.ts
